refactor out Points component
